Conflicts:
	src/Omma/AppBundle/Resources/views/Meeting/edit.html.twig

+++ src/Omma/AppBundle/Controller/MeetingAttendeeController.php
@@ -7,7 +7,9 @@ use FOS\RestBundle\Routing\ClassResourceInterface;
 use Omma\AppBundle\Entity\Attendee;
 use Omma\AppBundle\Entity\Meeting;
 use Omma\AppBundle\Form\Type\MeetingAttendeeForm;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 
 /**
  *
@@ -16,12 +18,21 @@ use Symfony\Component\HttpFoundation\Request;
  */
 class MeetingAttendeeController extends FOSRestController implements ClassResourceInterface
 {
+    /**
+     * @Security("is_granted('view', meeting)")
+     *
+     * @param Meeting $meeting
+     *
+     * @return \Omma\AppBundle\Entity\Attendee[]
+     */
     public function cgetAction(Meeting $meeting)
     {
         return $meeting->getAttendees();
     }
 
     /**
+     * @Security("is_granted('edit', meeting)")
+     *
      * @param Request $request
      * @param Meeting $meeting
      *
@@ -36,6 +47,8 @@ class MeetingAttendeeController extends FOSRestController implements ClassResour
     }
 
     /**
+     * @Security("is_granted('edit', meeting)")
+     *
      * @param Request  $request
      * @param Meeting  $meeting
      * @param Attendee $attendee
@@ -47,13 +60,32 @@ class MeetingAttendeeController extends FOSRestController implements ClassResour
         return $this->processForm($request, $attendee);
     }
 
+    /**
+     * @Security("is_granted('view', meeting)")
+     *
+     * @param Meeting  $meeting
+     * @param Attendee $attendee
+     *
+     * @return Attendee
+     */
     public function getAction(Meeting $meeting, Attendee $attendee)
     {
         return $attendee;
     }
 
+    /**
+     * @Security("is_granted('edit', meeting)")
+     *
+     * @param Meeting  $meeting
+     * @param Attendee $attendee
+     *
+     * @return \FOS\RestBundle\View\View
+     */
     public function deleteAction(Meeting $meeting, Attendee $attendee)
     {
+        if ($attendee->isOwner()) {
+            return $this->view("owner can't be removed", Response::HTTP_BAD_REQUEST);
+        }
         $meeting->removeAttendee($attendee);
         $this->get("omma.app.manager.attendee")->delete($attendee);
